增删改

+++ server/src/main/java/com/alimama/server/mapper/EmployeeMapper.java
@@ -29,7 +29,7 @@ public interface EmployeeMapper {
 
     int updateByPrimaryKeySelective(Employee record);
 
-    int updateByPrimaryKey(Employee record);
+    Long updateByPrimaryKey(Employee record);
 
     List<Employee> selectPageByExample(EmployeeExample employeeExample, @Param("pageNum")int page, @Param("pageSize")int limit);
